feat(커스텀타이머) : 커스텀타이머 생성 기능 개선 (#248)

+++ src/main/java/com/project200/undabang/timer/custom/dto/response/CustomTimerCreateResponse.java
@@ -1,4 +1,4 @@
 package com.project200.undabang.timer.custom.dto.response;
 
-public record CustomTimerCreateResponse(int customTimerId) {
+public record CustomTimerCreateResponse(Long customTimerId) {
 }
